+++ extension/chrome/settings/setup/setup-recover-key.ts
